+++ src/norm/sqlite.nim
@@ -28,173 +28,15 @@ Nim ``times.DateTime`` values are stored as ``INTEGER`` Unix epoch timestamps.
 
 
 import strutils, macros, typetraits, logging, options
-import ndb / sqlite
+import ndb/sqlite
 
-import rowutils, objutils, pragmas
+import sqlite/[rowutils, sqlgen], objutils, typedefutils
 
 
-export strutils, macros, logging, options
+export logging, options
 export sqlite
-export rowutils, objutils, pragmas
+export rowutils, sqlgen, objutils
 
-
-proc `$`*(query: SqlQuery): string = $ string query
-
-proc getTable*(objRepr: ObjRepr): string =
-  ##[ Get the name of the DB table for the given object representation:
-  ``table`` pragma value if it exists or lowercased type name otherwise.
-  ]##
-
-  result = objRepr.signature.name.toLowerAscii()
-
-  for prag in objRepr.signature.pragmas:
-    if prag.name == "table" and prag.kind == pkKval:
-      return $prag.value
-
-proc getTable*(T: typedesc): string =
-  ##[ Get the name of the DB table for the given type: ``table`` pragma value if it exists
-  or lowercased type name otherwise.
-  ]##
-
-  when T.hasCustomPragma(table): T.getCustomPragmaVal(table)
-  else: ($T).toLowerAscii()
-
-proc getColumn*(fieldRepr: FieldRepr): string =
-  ##[ Get the name of DB column for a field: ``dbCol`` pragma value if it exists
-  or field name otherwise.
-  ]##
-
-  result = fieldRepr.signature.name
-
-  for prag in fieldRepr.signature.pragmas:
-    if prag.name == "dbCol" and prag.kind == pkKval:
-      return $prag.value
-
-proc getColumns*(obj: object, force = false): seq[string] =
-  ## Get DB column names for an object as a sequence of strings.
-
-  for field, _ in obj.fieldPairs:
-    if force or not obj.dot(field).hasCustomPragma(ro):
-      when obj.dot(field).hasCustomPragma(dbCol):
-        result.add obj.dot(field).getCustomPragmaVal(dbCol)
-      else:
-        result.add field
-
-proc getDbType(fieldRepr: FieldRepr): string =
-  ## SQLite-specific mapping from Nim types to SQL data types.
-
-  for prag in fieldRepr.signature.pragmas:
-    if prag.name == "dbType" and prag.kind == pkKval:
-      return $prag.value
-
-  result =
-    if fieldRepr.typ.kind == nnkIdent:
-      case $fieldRepr.typ
-        of "int", "bool", "DateTime": "INTEGER NOT NULL"
-        of "string": "TEXT NOT NULL"
-        of "float": "REAL NOT NULL"
-        else: "TEXT NOT NULL"
-    elif fieldRepr.typ.kind == nnkBracketExpr and $fieldRepr.typ[0] == "Option":
-      case $fieldRepr.typ[1]
-        of "int", "bool", "DateTime": "INTEGER"
-        of "string": "TEXT"
-        of "float": "REAL"
-        else: "TEXT"
-    else: "TEXT NOT NULL"
-
-proc genColStmt(fieldRepr: FieldRepr, dbObjReprs: openArray[ObjRepr]): string =
-  ## Generate SQL column statement for a field representation.
-
-  result.add fieldRepr.getColumn()
-  result.add " "
-  result.add getDbType(fieldRepr)
-
-  for prag in fieldRepr.signature.pragmas:
-    if prag.name == "pk" and prag.kind == pkFlag:
-      result.add " PRIMARY KEY"
-    elif prag.name == "unique" and prag.kind == pkFlag:
-      result.add " UNIQUE"
-    elif prag.name == "notNull" and prag.kind == pkFlag:
-      result.add " NOT NULL"
-    elif prag.name == "check" and prag.kind == pkKval:
-      result.add " CHECK $#" % $prag.value
-    elif prag.name == "default" and prag.kind == pkKval:
-      result.add " DEFAULT $#" % $prag.value
-    elif prag.name == "fk" and prag.kind == pkKval:
-      expectKind(prag.value, {nnkIdent, nnkDotExpr})
-      result.add case prag.value.kind
-        of nnkIdent:
-          " REFERENCES $# (id)" % [dbObjReprs.getByName($prag.value).getTable()]
-        of nnkDotExpr:
-          " REFERENCES $# ($#)" % [dbObjReprs.getByName($prag.value[0]).getTable(), $prag.value[1]]
-        else: ""
-    elif prag.name == "onUpdate" and prag.kind == pkKval:
-      result.add " ON UPDATE $#" % $prag.value
-    elif prag.name == "onDelete" and prag.kind == pkKval:
-      result.add " ON DELETE $#" % $prag.value
-
-proc genTableSchema(dbObjRepr: ObjRepr, dbObjReprs: openArray[ObjRepr]): string =
-  ## Generate table schema for an object representation.
-
-  result.add "CREATE TABLE $# (\n" % dbObjRepr.getTable()
-
-  var columns: seq[string]
-
-  for field in dbObjRepr.fields:
-    columns.add "\t$#" % genColStmt(field, dbObjReprs)
-
-  result.add columns.join(",\n")
-  result.add "\n)"
-
-proc genTableSchemas*(dbObjReprs: openArray[ObjRepr]): seq[SqlQuery] =
-  ## Generate table schemas for a list of object representations.
-
-  for dbObjRepr in dbObjReprs:
-    result.add sql genTableSchema(dbObjRepr, dbObjReprs)
-
-proc genDropTableQueries*(dbObjReprs: seq[ObjRepr]): seq[SqlQuery] =
-  ## Generate ``DROP TABLE`` queries for a list of object representations.
-
-  for dbObjRepr in dbObjReprs:
-    result.add sql "DROP TABLE IF EXISTS $#" % dbObjRepr.getTable()
-
-proc genInsertQuery*(obj: object, force: bool): SqlQuery =
-  ## Generate ``INSERT`` query for an object.
-
-  let
-    fields = obj.getColumns(force)
-    placeholders = '?'.repeat(fields.len)
-
-  result = sql "INSERT INTO $# ($#) VALUES ($#)" % [type(obj).getTable(), fields.join(", "),
-                                                    placeholders.join(", ")]
-
-proc genGetOneQuery*(obj: object, condition: string): SqlQuery =
-  ## Generate ``SELECT`` query to fetch a single record for an object.
-
-  sql "SELECT $# FROM $# WHERE $#" % [obj.getColumns(force=true).join(", "),
-                                      type(obj).getTable(), condition]
-
-proc genGetManyQuery*(obj: object, condition: string): SqlQuery =
-  ## Generate ``SELECT`` query to fetch multiple records for an object.
-
-  sql "SELECT $# FROM $# WHERE $# LIMIT ? OFFSET ?" % [obj.getColumns(force=true).join(", "),
-                                                       type(obj).getTable(), condition]
-
-proc genUpdateQuery*(obj: object, force: bool): SqlQuery =
-  ## Generate ``UPDATE`` query for an object.
-
-  var fieldsWithPlaceholders: seq[string]
-
-  for field in obj.getColumns(force):
-    fieldsWithPlaceholders.add field & " = ?"
-
-  result = sql "UPDATE $# SET $# WHERE id = ?" % [type(obj).getTable(),
-                                                  fieldsWithPlaceholders.join(", ")]
-
-proc genDeleteQuery*(obj: object): SqlQuery =
-  ## Generate ``DELETE`` query for an object.
-
-  sql "DELETE FROM $# WHERE id = ?" % type(obj).getTable()
 
 template genWithDb(connection, user, password, database: string,
                    tableSchemas, dropTableQueries: openArray[SqlQuery]): untyped {.dirty.} =
@@ -367,30 +209,37 @@ template genWithDb(connection, user, password, database: string,
     withCustomDb(connection, user, password, database):
       body
 
+macro dbTypes*(body: untyped): untyped =
+  result = newStmtList()
 
-proc ensureIdFields(typeSection: NimNode): NimNode =
-  ## Check if ``id`` field is in the object definition, insert it if it's not.
+  for node in body:
+    expectKind(node, nnkTypeSection)
+    result.add ensureIdFields(node)
 
-  result = newNimNode(nnkTypeSection)
+macro dbFromTypes*(connection, user, password, database: string,
+                   types: openArray[typedesc]): untyped =
+  ##[ DB models definition. Models are defined as regular Nim objects in regular ``type`` sections.
 
-  for typeDef in typeSection:
-    var objRepr = typeDef.toObjRepr()
+  ``connection``, ``user``, ``password``, ``database`` are the same args accepted
+  by a standard ``dbConn`` instance.
+
+  ``types`` is a list of predefined types to create table schemas from.
+
+  The macro generates ``withDb`` template that wraps all DB interations.
+  ]##
+
+  var dbObjReprs: seq[ObjRepr]
+
+  for typ in types:
+    let objRepr = getImpl(typ).toObjRepr()
 
     if "id" notin objRepr.fieldNames:
-      let idField = FieldRepr(
-        signature: SignatureRepr(
-          name: "id",
-          exported: true,
-          pragmas: @[
-            PragmaRepr(name: "pk", kind: pkFlag),
-            PragmaRepr(name: "ro", kind: pkFlag)
-          ]
-        ),
-        typ: ident "int"
-      )
-      objRepr.fields.insert(idField, 0)
+      error "Type '$#' is missing 'id' field. Put it under 'ensureIdFields' macro." % $typ
 
-    result.add objRepr.toTypeDef()
+    dbObjReprs.add getImpl(typ).toObjRepr()
+
+  result = getAst genWithDb(connection, user, password, database,
+                            genTableSchemas(dbObjReprs), genDropTableQueries(dbObjReprs))
 
 macro db*(connection, user, password, database: string, body: untyped): untyped =
   ##[ DB models definition. Models are defined as regular Nim objects in regular ``type`` sections.
@@ -407,7 +256,7 @@ macro db*(connection, user, password, database: string, body: untyped): untyped 
 
   for node in body:
     if node.kind == nnkTypeSection:
-      let typeSection = node.ensureIdFields()
+      let typeSection = ensureIdFields(node)
 
       result.add typeSection