+++ tests/user.nim
@@ -1,0 +1,10 @@
+import times
+
+import norm/sqlite
+
+
+dbTypes:
+  type
+    User* {.table: "users".} = object
+      email*: string
+      lastLogin*: DateTime